backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-jose[cryptography]==3.3.0
firebase-admin==6.2.0
python-multipart==0.0.6
pydantic==1.10.13
python-dotenv==1.0.0
